More solid post rm

+++ src/main/resources/postrm.ftl
@@ -6,7 +6,9 @@ fi
 
 case "$1" in
         purge|remove)
-                rm -R "${config.installDir}"
+                if [ -d "${config.installDir}" ]; then
+                    rm -R "${config.installDir}"
+                fi
         ;;
         upgrade|failed-upgrade|abort-install|abort-upgrade|disappear)
         ;;
